requirements.txt
# Core
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
feedparser==6.0.10

# Data Processing
pandas==2.0.3
spacy==3.7.2
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.7.0/en_core_web_lg-3.7.0-py3-none-any.whl

# Security Tools
wafw00f==2.1.0
builtwith==1.3.3
python-whois==0.8.0

# APIs
slack-sdk==3.27.1
linkedin-api==2.0.0
pyhunter==1.4

# Utilities
pydantic==2.5.3
ratelimit==2.2.1
PyYAML>=5.0.0
python-dateutil>=2.8.0

# Google Sheets Integration
gspread==5.12.0
google-auth==2.27.0
